better management of wrong format in ext. formula files

+++ src/alfa_CR6_frontend/pages.py
@@ -834,22 +834,24 @@ class OrderPage(BaseStackedPage):
         order = app.create_order(path_to_file, n_of_jars=n)
 
         if order:
+            properties = json.loads(order.json_properties)
+            logging.warning(f"properties:{properties}")
+            if properties.get('meta') and not properties['meta'].get('error'):
+                args_to_print = sorted([[str(j.barcode), ] + j.extra_lines_to_print for j in order.jars])
+                logging.warning(f"file_name:{file_name}, args_to_print:{args_to_print}")
 
-            args_to_print = sorted([[str(j.barcode), ] + j.extra_lines_to_print for j in order.jars])
-            logging.warning(f"file_name:{file_name}, args_to_print:{args_to_print}")
+                def cb_():
+                    for a in args_to_print:
+                        logging.warning(f"a:{a}")
+                        response = dymo_print(*a)
+                        logging.warning(f"response:{response}")
+                        time.sleep(.05)
 
-            def cb_():
-                for a in args_to_print:
-                    logging.warning(f"a:{a}")
-                    response = dymo_print(*a)
-                    logging.warning(f"response:{response}")
-                    time.sleep(.05)
+                msg_ = tr_("confirm printing {} barcodes?").format(len(args_to_print))
+                self.main_window.open_input_dialog(message=msg_, content="{}".format(
+                    [l[0] for l in args_to_print]), ok_cb=cb_)
 
-            msg_ = tr_("confirm printing {} barcodes?").format(len(args_to_print))
-            self.main_window.open_input_dialog(message=msg_, content="{}".format(
-                [l[0] for l in args_to_print]), ok_cb=cb_)
-
-            model.remove_file(file_name)
+                model.remove_file(file_name)
 
         self.populate_file_table()
         self.populate_order_table()
